go brr

+++ GRAPH_SEARCH_MAP/GRAPH_SEARCH.h
@@ -26,7 +26,7 @@ State* createGoalState();
 void Print_State(const State *const state);
 void Print_Action(const enum ACTIONS action);
 int Result(const State *const parent_state, const enum ACTIONS action, Transition_Model *const trans_model);
-float Compute_Heuristic_Function(const State *const state, const State *const goal); 
+int Compute_Heuristic_Function(const State *const state, const State *const goal); 
 int Goal_Test(const State *const state, const State *const goal_state);
 int CountPassiveLights(const State *const state);
 // ======= YOU DO NOT NEED TO CHANGE THIS COMPULSORY DECLARATIONS EXCEPT INSERTION OF THE GENERALIZED A* ALGORITHM==============
